Did code cleanup, fixed structure of project
