+++ backend/api/misc/__init__.py
@@ -1,0 +1,2 @@
+from .custom_permissions import IsAdminOrReadOnly
+from .middlevare import SaveBrowsingHistoryMiddleware
