+++ backend/api/models/__init__.py
@@ -1,0 +1,3 @@
+from .exercise import Exercise
+from .system_user import SystemUser
+from .browsing_history import BrowsingHistory
