+++ backend/api/models/system_user.py
@@ -1,0 +1,16 @@
+from django.contrib.auth.models import AbstractUser, Group, Permission
+from django.db import models
+
+class SystemUser(AbstractUser):
+
+    email = models.EmailField(unique=True)
+    groups = models.ManyToManyField(
+        Group,
+        related_name="custom_user_set", 
+        blank=True,
+    )
+    user_permissions = models.ManyToManyField(
+        Permission,
+        related_name="custom_user_permissions_set",  
+        blank=True,
+    )
