+++ backend/api/serializers/__init__.py
@@ -1,0 +1,4 @@
+from .system_user_serializer import SystemUserSerializer
+from .exercise_serializer import ExerciseSerializer
+from .browsing_history_serializer import BrowsingHistorySerializer
+from .register_serializer import RegisterSerializer
