+++ backend/api/serializers/browsing_history_serializer.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from api.models import BrowsingHistory
+
+class BrowsingHistorySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = BrowsingHistory
+        fields = ['user', 'exercise', 'timestamp']
