+++ backend/api/serializers/exercise_serializer.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from api.models import Exercise
+
+class ExerciseSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Exercise
+        fields = ['id', 'name', 'muscle_group', 'difficulty', 'exercise_type', 'description', 'tutorial']
