+++ backend/api/serializers/register_serializer.py
@@ -1,0 +1,19 @@
+from rest_framework import serializers
+from api.models import SystemUser
+
+class RegisterSerializer(serializers.ModelSerializer):
+    password = serializers.CharField(write_only=True, required=True)
+    email = serializers.EmailField(required=True)
+    username = serializers.CharField(required=True)
+
+    class Meta:
+        model = SystemUser
+        fields = ['username', 'email', 'password']
+
+    def create(self, validated_data):
+        user = SystemUser.objects.create_user(
+            username=validated_data['username'],
+            email=validated_data['email'],
+            password=validated_data['password']
+        )
+        return user
